Refactor AI into separate module
