+++ index.html
@@ -94,6 +94,7 @@
     </div>
   </div>
 
+  <script src="js/ai.js"></script>
   <script src="js/core.js"></script>
 </body>
 </html>
